break `streams_and_futures` into two modules
